requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
Flask==2.0.2
gunicorn==20.1.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.1.0
MarkupSafe==2.0.1
nltk==3.6.5
npm==0.1.1
optional-django==0.1.0
Pillow==9.0.1
PyPDF2==1.26.0
requests==2.26.0
tqdm==4.62.3
urllib3==1.26.7
Werkzeug==2.0.2